Change float type variables to double.
We don't need float as we are not too worried about
space savings and we want to avoid rounding errors.

+++ src/main/java/seng201/team56/models/Rarity.java
@@ -8,15 +8,15 @@ import java.util.Random;
  * @author Sean Reitsma
  */
 public enum Rarity {
-    COMMON(10,20,1.5f,3, List.of(ResourceType.BCO_QUICHE,ResourceType.COQ_A_VIN)),
-    UNCOMMON(20, 40, 0.8f, 1.4f, List.of(ResourceType.BCO_QUICHE, ResourceType.COQ_A_VIN, ResourceType.NICOISE_SALAD)),
-    RARE(40, 80, 0.3f, 1, List.of(ResourceType.GRUYER_CHEESE_SOUFFLE, ResourceType.BOUILLABAISSE)),
-    EPIC(90, 110, 0.1f, 0.25f, List.of(ResourceType.CREME_BRULEE)),
-    LEGENDARY(150, 200, 0.001f, 0.1f, List.of(ResourceType.RATATOUILLE));
+    COMMON(10,20,1.5,3, List.of(ResourceType.BCO_QUICHE,ResourceType.COQ_A_VIN)),
+    UNCOMMON(20, 40, 0.8, 1.4, List.of(ResourceType.BCO_QUICHE, ResourceType.COQ_A_VIN, ResourceType.NICOISE_SALAD)),
+    RARE(40, 80, 0.3, 1, List.of(ResourceType.GRUYER_CHEESE_SOUFFLE, ResourceType.BOUILLABAISSE)),
+    EPIC(90, 110, 0.1, 0.25, List.of(ResourceType.CREME_BRULEE)),
+    LEGENDARY(150, 200, 0.001, 0.1, List.of(ResourceType.RATATOUILLE));
     private final int resourceAmountMin;
     private final int resourceAmountMax;
-    private final float speedMin;
-    private final float speedMax;
+    private final double speedMin;
+    private final double speedMax;
     private final List<ResourceType> types;
 
     /**
@@ -27,7 +27,7 @@ public enum Rarity {
      * @param speedMax Maximum speed (min duration)
      * @param types A list of possible {@link ResourceType}'s
      */
-    Rarity(int resourceAmountMin, int resourceAmountMax, float speedMin, float speedMax, List<ResourceType> types) {
+    Rarity(int resourceAmountMin, int resourceAmountMax, double speedMin, double speedMax, List<ResourceType> types) {
         this.resourceAmountMin = resourceAmountMin;
         this.resourceAmountMax = resourceAmountMax;
         this.speedMin = speedMin;
@@ -75,7 +75,7 @@ public enum Rarity {
      * Getter for speedMin
      * @return speedMin
      */
-    public float getSpeedMin() {
+    public double getSpeedMin() {
         return speedMin;
     }
 
@@ -83,7 +83,7 @@ public enum Rarity {
      * Getter for speedMax
      * @return speedMax
      */
-    public float getSpeedMax() {
+    public double getSpeedMax() {
         return speedMax;
     }
